*Should* rule for running CI using node
The reason being that if a purely SASS module uses ruby for CI, but then the module has some js added it won't be possible to use all your favourite js tools to compile and test your js. The inverse isn't true because testing sass (thus far) has consisted of compiling a sass file and checking the output - you can do this easily in node. Also a set of tests for sass written in js is going to be far more comprehensible to your average front-end dev than the equivalent in ruby.

*disclaimer - I don't know travis all that well, but a cursory look at the docs suggests to me that you have to pick a single engine.