Simplify the homepage design
This commit:
- introduces a new layout for the homepage
- reduces the number of clickable things on the homepage.
- introduces new wording for the introduction of Origami

Why
Talking to people across the FT about how they came to Origami, and how they
it to use, many of them talked about the information density of the site and the
difficulty they had, as newbies, in understanding what Origami was and how easy
or difficult it was to use. This commit restructures the first point of contact
to help these users get to the technical information they need and introduces
the core concepts of Origami (services, tools, modules, the spec) further down
the page in an ordered way.