fix markdown
I'm not sure why but all of our headings where there was no space
between the hash and the first letter of the heading are no longer parsed by
kramdown properly.

so `#title` doesn't work but `# title` does. I can't replicate this locally,
having tried updating Jekyll to v3, and checking the version of Kramdown i'm
using.

The problem was probably introduced when GitHub pages started using Jekyll 3,
which they did 4 days ago. But as I said, I can't replicate it. The not being
able to replicate thing is a problem because it means my dev env is different to
the live one, and it will therefore be easy to reintroduce this kind of md error
in the future without noticing.