Hide the overview link on the spec subnav
The link redirects to the home page which is a little confusing as it feels
like it should be an overview of the spec instead.
Have only commented out the links because there are links that were concealed
and I wasn't sure if they should be kept just in case.