Suggest a change to silent mode specification
Currently silent mode allows for the default of silent mode to be
either true or false. I'd like to propose that we make silent mode
true by default so that all modules that support silent styles are
consistent and can be included without the need to double-check
the documentation every time a module needs to be included.