Propose a new "supportContact" property
This introduces a new `supportContact` property in the origami.json
spec. This property is used to document support contact details of the
team that maintains a component or service. It also deprecates the use
of an email address in the `support` property.

I'm proposing this so that the support details of a component are more
strict and readily machine-readable - we need this for the repo data
service, which will use the information to differentiate between Origami
components that are maintained by the Origami team vs another team.